Port vsg mesh shading classes to cross vendor extension
