+++ include/vsg/core/ConstVisitor.h
@@ -319,7 +319,7 @@ namespace vsg
         virtual void apply(const WriteTimestamp&);
         virtual void apply(const CopyQueryPoolResults&);
 
-        // rtx classes
+        // mesh shading classes
         virtual void apply(const DrawMeshTasks&);
         virtual void apply(const DrawMeshTasksIndirect&);
         virtual void apply(const DrawMeshTasksIndirectCount&);
