+++ include/vsg/core/Visitor.h
@@ -319,7 +319,7 @@ namespace vsg
         virtual void apply(WriteTimestamp&);
         virtual void apply(CopyQueryPoolResults&);
 
-        // rtx classes
+        // mesh shading classes
         virtual void apply(DrawMeshTasks&);
         virtual void apply(DrawMeshTasksIndirect&);
         virtual void apply(DrawMeshTasksIndirectCount&);
