+++ src/vsg/CMakeLists.txt
@@ -140,7 +140,6 @@ set(SOURCES
     text/Text.cpp
     text/TextGroup.cpp
 
-
     threading/Affinity.cpp
     threading/OperationThreads.cpp
 
@@ -176,9 +175,9 @@ set(SOURCES
     raytracing/TopLevelAccelerationStructure.cpp
     raytracing/TraceRays.cpp
 
-    rtx/DrawMeshTasks.cpp
-    rtx/DrawMeshTasksIndirect.cpp
-    rtx/DrawMeshTasksIndirectCount.cpp
+    meshshaders/DrawMeshTasks.cpp
+    meshshaders/DrawMeshTasksIndirect.cpp
+    meshshaders/DrawMeshTasksIndirectCount.cpp
 
     ui/UIEvent.cpp
     ui/ApplicationEvent.cpp
